gemini: Update QCOM init shell script based on latest CAF changes
* Quick reference:
https://us.codeaurora.org/cgit/quic/la/device/qcom/common/log/rootdir/etc/init.qcom.sh?h=LA.HB.1.3.1&id=ca96d984d779253e18ee3026ee07312149af63b1
* Relevant patches:

    From: Shiyu Dong <shiyud@codeaurora.org>
    Date: Mon, 18 Apr 2016 12:45:50 -0700
    Subject: Copy MBN folder if modem does not have version info

    Continue with MBN folder copying if modem does not have ver_info.txt


    From: Sooraj Sasindran <ssasin@codeaurora.org>
    Date: Tue, 10 Nov 2015 23:57:03 -0800
    Subject: FR 30314: Do not copy mbns during every boot up

    Do not copy mbns for every boot up, instead copy
    only for first boot up or after a factory reset or
    after modem update

    CRs-Fixed: 938004

+++ rootdir/init.qcom.sh
@@ -94,9 +94,20 @@ echo 1 > /data/misc/radio/db_check_done
 #
 # Make modem config folder and copy firmware config to that folder for RIL
 #
-rm -rf /data/misc/radio/modem_config
-mkdir /data/misc/radio/modem_config
-chmod 770 /data/misc/radio/modem_config
-cp -r /firmware/image/modem_pr/mcfg/configs/* /data/misc/radio/modem_config
-chown -hR radio.radio /data/misc/radio/modem_config
+if [ -f /data/misc/radio/ver_info.txt ]; then
+    prev_version_info=`cat /data/misc/radio/ver_info.txt`
+else
+    prev_version_info=""
+fi
+
+cur_version_info=`cat /firmware/verinfo/ver_info.txt`
+if [ ! -f /firmware/verinfo/ver_info.txt -o "$prev_version_info" != "$cur_version_info" ]; then
+    rm -rf /data/misc/radio/modem_config
+    mkdir /data/misc/radio/modem_config
+    chmod 770 /data/misc/radio/modem_config
+    cp -r /firmware/image/modem_pr/mcfg/configs/* /data/misc/radio/modem_config
+    chown -hR radio.radio /data/misc/radio/modem_config
+    cp /firmware/verinfo/ver_info.txt /data/misc/radio/ver_info.txt
+    chown radio.radio /data/misc/radio/ver_info.txt
+fi
 echo 1 > /data/misc/radio/copy_complete
